refactored sentinel injector for specific use of rounding type issue
